+++ install/x64-Debug/include/MCC/reprojectionError.h
@@ -18,8 +18,8 @@ double computeReprojectionErrors(const std::vector<std::vector<cv::Point3f> >& o
 	std::vector<float>& perViewErrors,
 	bool fisheye);
 
-double computeReprojectionErrors(const std::vector<cv::Point3f>& objectPoints,
-	const std::vector<cv::Point2f>& imagePoints,
+double computeReprojectionErrors(const std::vector<cv::Point3d>& objectPoints,
+	const std::vector<cv::Point2d>& imagePoints,
 	const cv::Mat& rvecs,
 	const cv::Mat& tvecs,
 	const cv::Mat& cameraMatrix,
